worker/requirements.txt
pika==1.1.0
opencv-python
python-dotenv
numpy
Pillow